+++ apollo-client/src/test/java/com/ctrip/framework/apollo/spring/JavaConfigAnnotationTest.java
@@ -20,6 +20,7 @@ import com.ctrip.framework.apollo.Config;
 import com.ctrip.framework.apollo.ConfigChangeListener;
 import com.ctrip.framework.apollo.ConfigFileChangeListener;
 import com.ctrip.framework.apollo.core.ConfigConsts;
+import com.ctrip.framework.apollo.internals.SimpleConfig;
 import com.ctrip.framework.apollo.internals.YamlConfigFile;
 import com.ctrip.framework.apollo.model.ConfigChange;
 import com.ctrip.framework.apollo.model.ConfigChangeEvent;
@@ -32,7 +33,9 @@ import com.google.common.util.concurrent.SettableFuture;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.Properties;
+import java.util.Queue;
 import java.util.UUID;
+import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.TimeUnit;
 import org.junit.After;
 import org.junit.Test;
@@ -55,10 +58,12 @@ import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anySetOf;
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Matchers.eq;
+import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -350,6 +355,68 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     assertEquals(1, fxApolloConfigInterestedKeys.getAllValues().size());
 
     assertEquals(Collections.singletonList(Sets.newHashSet("anotherKey")), fxApolloConfigInterestedKeys.getAllValues());
+  }
+
+  @Test
+  public void testApolloConfigChangeListenerWithInterestedKeyPrefixes() {
+    Config applicationConfig = mock(Config.class);
+
+    mockConfig(ConfigConsts.NAMESPACE_APPLICATION, applicationConfig);
+
+    TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean bean = getBean(
+        TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean.class, AppConfig10.class);
+
+    final ArgumentCaptor<Set> interestedKeyPrefixesArgumentCaptor = ArgumentCaptor
+        .forClass(Set.class);
+
+    verify(applicationConfig, times(1))
+        .addChangeListener(any(ConfigChangeListener.class), anySetOf(String.class),
+            interestedKeyPrefixesArgumentCaptor.capture());
+
+    assertEquals(1, interestedKeyPrefixesArgumentCaptor.getAllValues().size());
+
+    Set<String> result = Sets.newHashSet();
+    for (Set<String> interestedKeyPrefixes : interestedKeyPrefixesArgumentCaptor.getAllValues()) {
+      result.addAll(interestedKeyPrefixes);
+    }
+    assertEquals(Sets.newHashSet("logging.level", "number"), result);
+  }
+
+  @Test
+  public void testApolloConfigChangeListenerWithInterestedKeyPrefixes_fire() {
+    // default mock, useless here
+    // just for speed up test without waiting
+    mockConfig(ConfigConsts.NAMESPACE_APPLICATION, mock(Config.class));
+
+    SimpleConfig simpleConfig = spy(
+        this.prepareConfig(
+            TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1.SPECIAL_NAMESPACE,
+            new Properties()));
+
+    mockConfig(TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1.SPECIAL_NAMESPACE,
+        simpleConfig);
+
+    TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1 bean = getBean(
+        TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1.class, AppConfig11.class);
+
+    verify(simpleConfig, atLeastOnce())
+        .addChangeListener(any(ConfigChangeListener.class), anySetOf(String.class),
+            anySetOf(String.class));
+
+    Properties properties = new Properties();
+    properties.put("logging.level.com", "debug");
+    properties.put("logging.level.root", "warn");
+    properties.put("number.value", "333");
+
+    // publish config change
+    simpleConfig.onRepositoryChange(
+        TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1.SPECIAL_NAMESPACE, properties);
+
+    // get event from bean
+    ConfigChangeEvent configChangeEvent = bean.getConfigChangeEvent();
+    Set<String> interestedChangedKeys = configChangeEvent.interestedChangedKeys();
+    assertEquals(Sets.newHashSet("logging.level.com", "logging.level.root", "number.value"),
+        interestedChangedKeys);
   }
 
   @Test
@@ -735,6 +802,24 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
     }
   }
 
+  @Configuration
+  @EnableApolloConfig
+  static class AppConfig10 {
+    @Bean
+    public TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean bean() {
+      return new TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean();
+    }
+  }
+
+  @Configuration
+  @EnableApolloConfig
+  static class AppConfig11 {
+    @Bean
+    public TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1 bean() {
+      return spy(new TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1());
+    }
+  }
+
   static class TestApolloConfigBean1 {
     @ApolloConfig
     private Config config;
@@ -847,6 +932,30 @@ public class JavaConfigAnnotationTest extends AbstractSpringIntegrationTest {
         interestedKeys = {"anotherKey"})
     private void anotherOnChange(ConfigChangeEvent changeEvent) {
 
+    }
+  }
+
+  private static class TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean {
+
+    @ApolloConfigChangeListener(interestedKeyPrefixes = {"number", "logging.level"})
+    private void onChange(ConfigChangeEvent changeEvent) {
+    }
+  }
+
+  private static class TestApolloConfigChangeListenerWithInterestedKeyPrefixesBean1 {
+
+    static final String SPECIAL_NAMESPACE = "special-namespace-2021";
+
+    private final Queue<ConfigChangeEvent> configChangeEventQueue = new ArrayBlockingQueue<>(100);
+
+    @ApolloConfigChangeListener(value = SPECIAL_NAMESPACE, interestedKeyPrefixes = {"number",
+        "logging.level"})
+    private void onChange(ConfigChangeEvent changeEvent) {
+      this.configChangeEventQueue.add(changeEvent);
+    }
+
+    public ConfigChangeEvent getConfigChangeEvent() {
+      return this.configChangeEventQueue.poll();
     }
   }